refactoring linearization into its own module
